requirements.txt
# Core & Data
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
statsmodels>=0.13.0
yfinance

# Deep Learning & Math
torch>=1.10.0
arch>=5.0
scikit-learn

# Web / API
flask
gunicorn
pandas-gbq

# Cloud & Infra
google-cloud-bigquery
google-cloud-storage
google-cloud-aiplatform>=1.38.0

# Development & Testing
pytest
tqdm
python-dotenv